requirements.txt
ultralytics
pillow